+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Rename `fix-vulnerabilities` to `remediate-vulnerabilities`.
+
 ## [6.26.3] - 2024-05-21
 
 ### Fixed